Update sql_queries.py
The INSERT statements are mostly correct. Great job dealing with the DUPLICATE values using ON CONFLICT clauses.
However, using DO NOTHING is not always the best approach. Even though there's no correct answer here, sometimes we don't want to lose all values.

For the users table, this is especially important. The level column has information about the user membership (free/paid).

Please update your ON CONFLICT clause for this table and update the column level in case of conflicts.
If you want to learn how to deal with conflicts, check this link